move all segmeta reading and writing to a new module
