+++ pkg/es/reader/esQueryHandler.go
@@ -196,7 +196,7 @@ func getIndexNameAggOnly(aggName string, myid uint64) *structs.NodeResult {
 	totalHits := uint64(0)
 	bucketResults := make([]*structs.BucketResult, 0)
 
-	allSegmetas := segwriter.ReadAllSegmetas()
+	allSegmetas := segwriter.ReadGlobalSegmetas()
 
 	allCnts := segwriter.GetVTableCountsForAll(myid, allSegmetas)
 	segwriter.GetUnrotatedVTableCountsForAll(myid, allCnts)
