+++ pkg/segment/writer/segwriter_test.go
@@ -115,107 +115,6 @@ func Test_getActiveBaseSegDir(t *testing.T) {
 	assert.EqualValues(t, dataPath+"/"+config.GetHostID()+"/final/"+virtualTableName+"/"+streamid+"/1/", basedir)
 }
 
-func Test_ReplaceSingleSegMeta(t *testing.T) {
-	config.InitializeDefaultConfig(t.TempDir())
-	initSmr()
-
-	segMetasArr := ReadLocalSegmeta()
-	segMetas := make(map[string]*structs.SegMeta)
-	for _, smentry := range segMetasArr {
-		if smentry.SegmentKey == "key1" {
-			segMetas[smentry.SegmentKey] = smentry
-			break
-		}
-	}
-	assert.Empty(t, segMetas)
-
-	segMetaV1 := structs.SegMeta{
-		SegmentKey:  "key1",
-		RecordCount: 20,
-	}
-
-	AddOrReplaceRotatedSegment(segMetaV1)
-
-	segMetasArr = ReadLocalSegmeta()
-	segMetas = make(map[string]*structs.SegMeta)
-	for _, smentry := range segMetasArr {
-		if smentry.SegmentKey == "key1" {
-			segMetas[smentry.SegmentKey] = smentry
-			break
-		}
-	}
-
-	assert.Len(t, segMetas, 1)
-	assert.Equal(t, segMetaV1.SegmentKey, segMetas["key1"].SegmentKey)
-	assert.Equal(t, segMetaV1.RecordCount, segMetas["key1"].RecordCount)
-
-	segMetaV2 := structs.SegMeta{
-		SegmentKey:  "key1",
-		RecordCount: 50,
-	}
-
-	AddOrReplaceRotatedSegment(segMetaV2)
-
-	segMetasArr = ReadLocalSegmeta()
-	segMetas = make(map[string]*structs.SegMeta)
-	for _, smentry := range segMetasArr {
-		if smentry.SegmentKey == "key1" {
-			segMetas[smentry.SegmentKey] = smentry
-			break
-		}
-	}
-
-	assert.Len(t, segMetas, 1)
-	assert.Equal(t, segMetaV2.SegmentKey, segMetas["key1"].SegmentKey)
-	assert.Equal(t, segMetaV2.RecordCount, segMetas["key1"].RecordCount)
-}
-
-func Test_ReplaceMiddleSegMeta(t *testing.T) {
-	config.InitializeDefaultConfig(t.TempDir())
-	initSmr()
-
-	segMeta1 := structs.SegMeta{
-		SegmentKey:  "key1",
-		RecordCount: 20,
-	}
-	segMeta2V1 := structs.SegMeta{
-		SegmentKey:  "key2",
-		RecordCount: 50,
-	}
-	segMeta3 := structs.SegMeta{
-		SegmentKey:  "key3",
-		RecordCount: 100,
-	}
-
-	AddOrReplaceRotatedSegment(segMeta1)
-	AddOrReplaceRotatedSegment(segMeta2V1)
-	AddOrReplaceRotatedSegment(segMeta3)
-
-	segMeta2V2 := structs.SegMeta{
-		SegmentKey:  "key2",
-		RecordCount: 80,
-	}
-	AddOrReplaceRotatedSegment(segMeta2V2)
-
-	segMetasArr := ReadLocalSegmeta()
-	segMetas := make(map[string]*structs.SegMeta)
-	for _, smentry := range segMetasArr {
-		if smentry.SegmentKey == "key1" ||
-			smentry.SegmentKey == "key2" ||
-			smentry.SegmentKey == "key3" {
-			segMetas[smentry.SegmentKey] = smentry
-		}
-	}
-
-	assert.Len(t, segMetas, 3)
-	assert.Equal(t, segMeta1.SegmentKey, segMetas["key1"].SegmentKey)
-	assert.Equal(t, segMeta1.RecordCount, segMetas["key1"].RecordCount)
-	assert.Equal(t, segMeta2V2.SegmentKey, segMetas["key2"].SegmentKey)
-	assert.Equal(t, segMeta2V2.RecordCount, segMetas["key2"].RecordCount)
-	assert.Equal(t, segMeta3.SegmentKey, segMetas["key3"].SegmentKey)
-	assert.Equal(t, segMeta3.RecordCount, segMetas["key3"].RecordCount)
-}
-
 func Test_getNumberTypeAndVal(t *testing.T) {
 	cases := []struct {
 		input   string
